Use PublicAddress to override outgoing PORT IP
Normally, when the proxy sends a PORT command, it uses it’s listen
address as the IP. However, in a system where the public IP is actually
mapped to a private ip (Such as Amazon AWS), this port will be
incorrect.  I’ve added an optional PublicAddress config option to
override the IP Address sent by outgoing PORT commands. This option will
allow ftp-proxy to advertise the correct public IP address to servers
when it sends PORT commands.